Fix crash on Ruby 3.1 involving a change to the Psych gem.
