+++ lib/phobos/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module Phobos
-  VERSION = '2.1.0'
+  VERSION = '2.1.1'
 end
